# docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "用户中心服务 API 文档",
        "title": "User Hub API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {},
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8081",
    "paths": {
...
        "/api/v1/user-hub/identities": {
            "post": {
                "description": "用户或管理员为指定用户绑定一种新的登录方式（如新的账号密码、关联社交账号等）。",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "身份管理 (Identity Management)"
                ],
                "summary": "创建新身份",
                "parameters": [
                    {
                        "description": "创建身份请求的详细信息，包括用户ID、身份类型、标识符和凭证",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.CreateIdentityDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "身份创建成功，返回新创建的身份信息",
                        "schema": {
                            "$ref": "#/definitions/response.APIResponse-vo_IdentityVO"
                        }
                    },
                    "400": {
                        "description": "请求参数无效 (如JSON格式错误、必填项缺失) 或 业务逻辑错误 (如身份标识已存在)",
                        "schema": {
                            "$ref": "#/definitions/response.APIResponse-string"
                        }
                    },
                    "500": {
                        "description": "系统内部错误 (如数据库操作失败、密码加密失败)",
                        "schema": {
                            "$ref": "#/definitions/response.APIResponse-string"
                        }
                    }
                }
            }
        },
        "/api/v1/user-hub/identities/{identityID}": {
            "put": {
                "description": "用户或管理员修改指定身份ID的凭证信息（例如，重置密码）。",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "身份管理 (Identity Management)"
                ],
                "summary": "更新身份信息",
                "parameters": [
                    {
                        "type": "integer",
                        "format": "uint",
                        "description": "要更新的身份记录的唯一ID",
                        "name": "identityID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "更新身份请求的详细信息，主要包含新的凭证",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.UpdateIdentityDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "身份信息更新成功，返回更新后的身份信息",
                        "schema": {
                            "$ref": "#/definitions/response.APIResponse-vo_IdentityVO"
                        }
                    },
                    "400": {
                        "description": "请求参数无效 (如JSON格式错误、身份ID格式无效、新凭证无效)",
                        "schema": {
                            "$ref": "#/definitions/response.APIResponse-string"
                        }
                    },
                    "404": {
                        "description": "指定的身份记录不存在",
                        "schema": {
                            "$ref": "#/definitions/response.APIResponse-string"
                        }
                    },
                    "500": {
                        "description": "系统内部错误 (如数据库操作失败、密码加密失败)",
                        "schema": {
                            "$ref": "#/definitions/response.APIResponse-string"
                        }
                    }
                }
            },
            "delete": {
                "description": "用户或管理员注销或移除某个特定的登录方式（身份记录）。",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "身份管理 (Identity Management)"
                ],
                "summary": "删除身份",
                "parameters": [
                    {
                        "type": "integer",
                        "format": "uint",
                        "description": "要删除的身份记录的唯一ID",
                        "name": "identityID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "身份删除成功",
                        "schema": {
                            "$ref": "#/definitions/response.APIResponse-vo_Empty"
                        }
                    },
                    "400": {
                        "description": "请求参数无效 (如身份ID格式无效)",
                        "schema": {
                            "$ref": "#/definitions/response.APIResponse-string"
                        }
                    },
                    "404": {
                        "description": "指定的身份记录不存在 (如果服务层认为删除不存在的记录是错误)",
                        "schema": {
                            "$ref": "#/definitions/response.APIResponse-string"
                        }
                    },
                    "500": {
                        "description": "系统内部错误 (如数据库操作失败)",
                        "schema": {
                            "$ref": "#/definitions/response.APIResponse-string"
                        }
                    }
                }
            }
        },
...
        "/api/v1/user-hub/users/{userID}/identities": {
            "get": {
                "description": "管理员或用户本人查看指定用户ID关联的所有登录方式/身份凭证信息（不含敏感凭证内容）。",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "身份管理 (Identity Management)"
                ],
                "summary": "获取用户的所有身份信息",
                "parameters": [
                    {
                        "type": "string",
                        "description": "要查询的用户ID",
                        "name": "userID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "获取用户身份列表成功",
                        "schema": {
                            "$ref": "#/definitions/response.APIResponse-vo_IdentityList"
                        }
                    },
                    "400": {
                        "description": "请求参数无效 (如用户ID为空)",
                        "schema": {
                            "$ref": "#/definitions/response.APIResponse-string"
                        }
                    },
                    "404": {
                        "description": "指定的用户不存在 (如果服务层检查用户存在性)",
                        "schema": {
                            "$ref": "#/definitions/response.APIResponse-string"
                        }
                    },
                    "500": {
                        "description": "系统内部错误 (如数据库查询失败)",
                        "schema": {
                            "$ref": "#/definitions/response.APIResponse-string"
                        }
                    }
                }
            }
        },
        "/api/v1/user-hub/users/{userID}/identity-types": {
            "get": {
                "description": "用户或系统查看指定用户ID已绑定的所有登录方式的类型列表。",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "身份管理 (Identity Management)"
                ],
                "summary": "获取用户的所有身份类型",
                "parameters": [
                    {
                        "type": "string",
                        "description": "要查询的用户ID",
                        "name": "userID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "获取用户身份类型列表成功",
                        "schema": {
                            "$ref": "#/definitions/response.APIResponse-vo_IdentityTypeList"
                        }
                    },
                    "400": {
                        "description": "请求参数无效 (如用户ID为空)",
                        "schema": {
                            "$ref": "#/definitions/response.APIResponse-string"
                        }
                    },
                    "404": {
                        "description": "指定的用户不存在 (如果服务层检查用户存在性)",
                        "schema": {
                            "$ref": "#/definitions/response.APIResponse-string"
                        }
                    },
                    "500": {
                        "description": "系统内部错误 (如数据库查询失败)",
                        "schema": {
                            "$ref": "#/definitions/response.APIResponse-string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "dto.CreateIdentityDTO": {
            "type": "object",
            "required": [
                "credential",
                "identifier",
                "identity_type",
                "user_id"
            ],
            "properties": {
                "credential": {
                    "description": "凭证（如密码哈希、UnionID）",
                    "type": "string",
                    "example": "hashed_password"
                },
                "identifier": {
                    "description": "标识符（如账号、OpenID、手机号）",
                    "type": "string",
                    "example": "user123"
                },
                "identity_type": {
                    "description": "身份类型（0=账号密码, 1=小程序, 2=手机号）",
                    "allOf": [
                        {
                            "$ref": "#/definitions/enums.IdentityType"
                        }
                    ],
                    "example": 0
                },
                "user_id": {
                    "description": "用户 ID",
                    "type": "string",
                    "example": "123e4567-e89b-12d3-a456-426614174000"
                }
            }
        },
...
        "dto.UpdateIdentityDTO": {
            "type": "object",
            "required": [
                "credential"
            ],
            "properties": {
                "credential": {
                    "description": "新凭证（如新密码哈希）",
                    "type": "string",
                    "example": "new_hashed_password"
                }
            }
        },
...
        "enums.IdentityType": {
            "type": "integer",
            "enum": [
                0,
                1,
                2
            ],
            "x-enum-comments": {
                "AccountPassword": "账号密码（网站）",
                "Phone": "手机号（APP）",
                "WechatMiniProgram": "微信（小程序）"
            },
            "x-enum-varnames": [
                "AccountPassword",
                "WechatMiniProgram",
                "Phone"
            ]
        },
...
        "response.APIResponse-string": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "响应状态码，0 表示成功，其他值表示错误",
                    "type": "integer",
                    "example": 0
                },
                "data": {
                    "description": "响应数据，类型由 T 指定，若无数据则为 nil，且在 JSON 中省略",
                    "type": "string"
                },
                "message": {
                    "description": "可选的响应消息，若为空则不输出",
                    "type": "string",
                    "example": "success"
                }
            }
        },
        "response.APIResponse-vo_Empty": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "响应状态码，0 表示成功，其他值表示错误",
                    "type": "integer",
                    "example": 0
                },
                "data": {
                    "description": "响应数据，类型由 T 指定，若无数据则为 nil，且在 JSON 中省略",
                    "allOf": [
                        {
                            "$ref": "#/definitions/vo.Empty"
                        }
                    ]
                },
                "message": {
                    "description": "可选的响应消息，若为空则不输出",
                    "type": "string",
                    "example": "success"
                }
            }
        },
        "response.APIResponse-vo_IdentityList": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "响应状态码，0 表示成功，其他值表示错误",
                    "type": "integer",
                    "example": 0
                },
                "data": {
                    "description": "响应数据，类型由 T 指定，若无数据则为 nil，且在 JSON 中省略",
                    "allOf": [
                        {
                            "$ref": "#/definitions/vo.IdentityList"
                        }
                    ]
                },
                "message": {
                    "description": "可选的响应消息，若为空则不输出",
                    "type": "string",
                    "example": "success"
                }
            }
        },
        "response.APIResponse-vo_IdentityTypeList": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "响应状态码，0 表示成功，其他值表示错误",
                    "type": "integer",
                    "example": 0
                },
                "data": {
                    "description": "响应数据，类型由 T 指定，若无数据则为 nil，且在 JSON 中省略",
                    "allOf": [
                        {
                            "$ref": "#/definitions/vo.IdentityTypeList"
                        }
                    ]
                },
                "message": {
                    "description": "可选的响应消息，若为空则不输出",
                    "type": "string",
                    "example": "success"
                }
            }
        },
        "response.APIResponse-vo_IdentityVO": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "响应状态码，0 表示成功，其他值表示错误",
                    "type": "integer",
                    "example": 0
                },
                "data": {
                    "description": "响应数据，类型由 T 指定，若无数据则为 nil，且在 JSON 中省略",
                    "allOf": [
                        {
                            "$ref": "#/definitions/vo.IdentityVO"
                        }
                    ]
                },
                "message": {
                    "description": "可选的响应消息，若为空则不输出",
                    "type": "string",
                    "example": "success"
                }
            }
        },
        "vo.Empty": {
            "type": "object"
        },
        "vo.IdentityList": {
            "type": "object",
            "properties": {
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/vo.IdentityVO"
                    }
                }
            }
        },
        "vo.IdentityTypeList": {
            "type": "object",
            "properties": {
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/enums.IdentityType"
                    }
                }
            }
        },
        "vo.IdentityVO": {
            "type": "object",
            "properties": {
                "created_at": {
                    "description": "创建时间",
                    "type": "string",
                    "example": "2023-01-01T00:00:00Z"
                },
                "identifier": {
                    "description": "标识符（如账号、OpenID、手机号）",
                    "type": "string",
                    "example": "user123"
                },
                "identity_id": {
                    "description": "身份 ID",
                    "type": "integer",
                    "example": 1
                },
                "identity_type": {
                    "description": "身份类型（0=账号密码, 1=小程序, 2=手机号）",
                    "allOf": [
                        {
                            "$ref": "#/definitions/enums.IdentityType"
                        }
                    ],
                    "example": 0
                },
                "updated_at": {
                    "description": "更新时间",
                    "type": "string",
                    "example": "2023-01-01T00:00:00Z"
                },
                "user_id": {
                    "description": "用户 ID",
                    "type": "string",
                    "example": "123e4567-e89b-12d3-a456-426614174000"
                }
            }
        },
...
    }
}